feat: add support for chisel release format "v1"
This commits adds support for "v1" which changes the "v1-public-keys"
field into "public-keys". The "chisel-v1" format is still supported.
